add benchmark & clean repo

+++ src/custom_envs.py
@@ -882,7 +882,7 @@ class CustomRotaryInvertedPendulumEnv(InvertedDoublePendulumEnv):
                 np.clip(self.data.qvel, -10, 10),
                 np.clip(self.data.qfrc_constraint, -10, 10)[:1],
             ]
-        ).ravel()
+        , dtype=np.float32).ravel()
 
     def compute_reward(self, x, y, terminated):
         target_pos = np.array([0, 0, 0.5365])
